Search matching brace
The keybinding for searching matching braces is only useful, when the cursor is located close to a brace. If the cursor is in the middle of a block the keybinding is useless. I've changed the behaviour in order to go to the start of the current block. It is given by a solitary opening brace.