Select to matching brace
The feature of jumping in between braces is really useful. Often I
find me wanting to select the text in between braces. So I added the
feature to geany.

The default keybinding I used is Ctrl-Shift-B because Ctrl-B goes
to the matching brace.

geany.txt is updated for documentation.

Conflicts:

	src/editor.c